--- a/tests/test_routes.py
+++ b/tests/test_routes.py
@@ -166,7 +166,9 @@ def test_matching(path, headers, expected):
     request = Request(path, headers, None, "GET", None, app)
 
     try:
-        app.router.get(request=request)
+        app.router.get(
+            request.path, request.method, request.headers.get("host")
+        )
     except NotFound:
         response = 404
     except Exception: