@ahopkins 2021-03-21
Agreed. I think a large focus for the next three months is to scrutinize every line that is executed in the request/response cycle.